Move remaining provider scheduling utils into @medplum/core
The scheduling extension URIs and transient identifier helper are part of the
Medplum scheduling data model rather than provider-specific UI code, so they
belong alongside SchedulingParametersURI in core where server-side callers can
reach them. Also adds coverage for SchedulingTransientIdentifier.remove, which
had none.
