--- examples/medplum-provider/src/utils/scheduling.ts
@@ -1,23 +1,0 @@
-// SPDX-FileCopyrightText: Copyright Orangebot, Inc. and Medplum contributors
-// SPDX-License-Identifier: Apache-2.0
-import { generateId, getIdentifier, setIdentifier } from '@medplum/core';
-import type { Identifier, Resource } from '@medplum/fhirtypes';
-const MedplumSchedulingTransientIdentifierURI = 'https://medplum.com/fhir/scheduling-transient-id';
-export const SchedulingEncounterCodingURI = 'https://medplum.com/fhir/StructureDefinition/SchedulingEncounterCoding';
-export const SchedulingPlanDefinitionURI = 'https://medplum.com/fhir/StructureDefinition/SchedulingPlanDefinition';
-
-export const SchedulingTransientIdentifier = {
-  set(resource: Resource & { identifier?: Identifier[] }) {
-    setIdentifier(resource, MedplumSchedulingTransientIdentifierURI, generateId(), { use: 'temp' });
-  },
-
-  get(resource: Resource) {
-    return getIdentifier(resource, MedplumSchedulingTransientIdentifierURI);
-  },
-
-  remove(resource: Resource & { identifier?: Identifier[] }) {
-    resource.identifier = resource.identifier?.filter(
-      (identifier) => identifier.system !== MedplumSchedulingTransientIdentifierURI
-    );
-  },
-};
